Adjustments and created new ChatTest class
