+++ src/test/java/at/aau/serg/websocketdemoserver/model/game/ChatTest.java
@@ -1,0 +1,55 @@
+package at.aau.serg.websocketdemoserver.model.game;
+
+import at.aau.serg.websocketdemoserver.msg.ChatMessage;
+import org.testng.annotations.Test;
+
+import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class ChatTest {
+
+
+    @Test
+    public void testConstructorAndGetters() {
+        String sender = "Daniel";
+        String text = "Hello, Daniel!";
+        String roomID = "room123";
+        String msgIdentifier = "msg456";
+
+        ChatMessage message = new ChatMessage(sender, text, roomID, msgIdentifier);
+
+        assertEquals(sender, message.getSender());
+        assertEquals(text, message.getText());
+        assertEquals(roomID, message.getRoomID());
+        assertEquals(msgIdentifier, message.getMsgIdentifier());
+    }
+
+    @Test
+    public void testSetters() {
+        ChatMessage message = new ChatMessage();
+
+        String sender = "Daniel";
+        String text = "Hello, Daniel!";
+        String roomID = "room123";
+        String msgIdentifier = "msg456";
+
+        message.setSender(sender);
+        message.setText(text);
+        message.setRoomID(roomID);
+        message.setMsgIdentifier(msgIdentifier);
+
+        assertEquals(sender, message.getSender());
+        assertEquals(text, message.getText());
+        assertEquals(roomID, message.getRoomID());
+        assertEquals(msgIdentifier, message.getMsgIdentifier());
+    }
+    @Test
+    public void testDefaultConstructor() {
+        ChatMessage message = new ChatMessage();
+
+        assertNull(message.getSender());
+        assertNull(message.getText());
+        assertNull(message.getRoomID());
+        assertNull(message.getMsgIdentifier());
+    }
+}
