+++ src/test/java/at/aau/serg/websocketdemoserver/model/game/GameboardTest.java
@@ -34,38 +34,4 @@ public class GameboardTest {
             assertEquals(maxMolehills, molehillCount);
         }
 
-    @Test
-    public void testConstructorAndGetters() {
-        String sender = "Daniel";
-        String text = "Hello, Daniel!";
-        String roomID = "room123";
-        String msgIdentifier = "msg456";
-
-        ChatMessage message = new ChatMessage(sender, text, roomID, msgIdentifier);
-
-        assertEquals(sender, message.getSender());
-        assertEquals(text, message.getText());
-        assertEquals(roomID, message.getRoomID());
-        assertEquals(msgIdentifier, message.getMsgIdentifier());
-    }
-
-    @Test
-    public void testSetters() {
-        ChatMessage message = new ChatMessage();
-
-        String sender = "Daniel";
-        String text = "Hello, Daniel!";
-        String roomID = "room123";
-        String msgIdentifier = "msg456";
-
-        message.setSender(sender);
-        message.setText(text);
-        message.setRoomID(roomID);
-        message.setMsgIdentifier(msgIdentifier);
-
-        assertEquals(sender, message.getSender());
-        assertEquals(text, message.getText());
-        assertEquals(roomID, message.getRoomID());
-        assertEquals(msgIdentifier, message.getMsgIdentifier());
-    }
     }
